--- .env.EXAMPLE
@@ -1,3 +1,0 @@
-DB_NAME='crowdfund_db'
-DB_USER=''
-DB_PASSWORD=''
