+++ db/schema.sql
@@ -1,2 +1,2 @@
-DROP DATABASE IF EXISTS crowdfund_db;
-CREATE DATABASE crowdfund_db;
+DROP DATABASE IF EXISTS projectSpot_db;
+CREATE DATABASE projectSpot_db;
