+++ public/css/style.css
@@ -1,7 +1,3 @@
-header, .project:not(:last-child) {
-  border-bottom: 1px solid #ccc;
-}
-
 .no-button {
   background: none;
   border: none;
@@ -22,3 +18,4 @@ header, .project:not(:last-child) {
   font-weight: bold;
   margin-left: -5px;
 }
+
