+++ public/seeds/projectData.json
@@ -1,0 +1,17 @@
+[
+  {
+    "name": "Music Near Me",
+    "description": "A mobile app that will send you notifications whenever a concert is playing in your area.",
+    "needed_funding": 5000
+  },
+  {
+    "name": "The Ultimate Tech Quiz",
+    "description": "A web app that will give users 10 new technical questions each day and track their progress in things like programming, cybersecurity, database architecture, and more!",
+    "needed_funding": 10000
+  },
+  {
+    "name": "Roll 'Em Up",
+    "description": "A game for Windows and macOS where players move a ball through a series of increasingly challenging mazes.",
+    "needed_funding": 800
+  }
+]
