+++ views/homepage.handlebars
@@ -1,24 +1,27 @@
 <!-- Loop over each object in the projects array -->
-{{#each projects as |project| }}
-<div class="row mb-4 project">
-  <div class="col-md-5">
-    <h2>
-      <!-- Call helper function with triple brackets to render the emoji span as HTML -->
-      {{{get_emoji}}}
-      <a href="/project/{{project.id}}">{{project.name}}</a>
-    </h2>
-    <p>
-      <span for="img" aria-label="money">💲</span>
-      <!-- Pass needed_funding value to the helper function -->
-      <span class="dollar-amount">{{format_amount project.needed_funding}}</span>
-      needed!
-    </p>
-    <p>Created by {{project.user.name}} on {{format_date project.date_created}}</p>
+{{#each projects as |project|}}
+  <div class="row mb-4 project">
+    <div class="card" style="width: 18rem;">
+      <img src="..." class="card-img-top" alt="..." />
+      <div class="card-body">
+        <h5 class="card-title">
+          {{project.name}}</h5>
+        <p class="card-description">{{project.description}}</p>
+      </div>
+      <ul class="list-group list-group-flush">
+        <li class="list-group-item">{{format_date project.date_created}}</li>
+        <li class="list-group-item"><span class="dollar-amount">{{format_amount
+              project.needed_funding
+            }}</span>
+          needed!</li>
+        <li class="list-group-item">Created by
+          {{project.user.name}}</li>
+      </ul>
+      <div class="card-body">
+        <a href="#" class="card-link"><a href="/project/{{project.id}}">Card
+            link</a></a>
+        <a href="#" class="card-link">Another link</a>
+      </div>
+    </div>
   </div>
-  <div class="col-md-7">
-    <p>
-      {{project.description}}
-    </p>
-  </div>
-</div>
 {{/each}}
