+++ views/signup.handlebars
@@ -1,0 +1,76 @@
+<div class="container">
+        <div class="row">
+			<div class="col-md-5 mx-auto">
+			<div id="first">
+				<div class="myform form ">
+					 <div class="logo mb-3">
+						 <div class="col-md-12 text-center">
+							<h1>Login</h1>
+						 </div>
+					</div>
+                   <form class="form login-form">
+                           <div class="form-group">
+                              <label for="email-login">Email address</label>
+                              <input type="email" class="form-input" id="email-login" placeholder="Enter email">
+                           </div>
+                           <div class="form-group">
+                              <label for="password-login">Password</label>
+                              <input type="password" id="password"  class="form-input" placeholder="Enter Password">
+                           </div>
+                           <div class="form-group">
+                              <p class="text-center">By signing up you accept our <a href="./terms.handlebars">Terms Of Use</a></p>
+                           </div>
+                           <div class="col-md-12 text-center ">
+                              <button type="submit" class=" btn btn-block mybtn btn-primary tx-tfm">Login</button>
+                           </div>
+                           <div class="col-md-12 ">
+                              <div class="login-or">
+                                 <hr class="hr-or">
+                                 <span class="span-or">or</span>
+                              </div>
+                           </div>
+                           <div class="col-md-12 mb-3">
+                              <p class="text-center">
+                                 <a href="javascript:void();" class="google btn mybtn"><i class="fa fa-google-plus">
+                                 </i> Signup using Google
+                                 </a>
+                              </p>
+                           </div>
+                           <div class="form-group">
+                              <p class="text-center">Don't have account? <a href="./login.handlebars" id="signup">Sign up here</a></p>
+                           </div>
+                        </form>
+                        </div>
+                           <div class="myform form ">
+                        <div class="logo mb-3">
+                           <div class="col-md-12 text-center">
+                              <h1 >Signup</h1>
+                           </div>
+                        </div>
+                        <form action="#" name="registration">
+                           <div class="form-group">
+                              <label for="name-signup">Enter your Name</label>
+                              <input type="text" class="form-input" id="name-signup" placeholder="Enter you name...">
+                           </div>
+                           <div class="form-group">
+                              <label for="email-login">Email address</label>
+                              <input type="email" class="form-input" id="email-signup" placeholder="Enter your email">
+                           </div>
+                           <div class="form-group">
+                              <label for="exampleInputEmail1">Password</label>
+                              <input class="form-input" type="password" id="password-signup" placeholder="Enter your password" />
+                           </div>
+                           <div class="col-md-12 text-center mb-3">
+                              <button type="submit" class=" btn btn-block mybtn btn-primary tx-tfm">Get Started For Free</button>
+                           </div>
+                           <div class="col-md-12 ">
+                              <div class="form-group">
+                                 <p class="text-center"><a href="./login" id="signin">Already have an account?</a></p>
+                              </div>
+                           </div>
+                            </div>
+                        </form>
+                     </div>     
+	    </div>
+  	</div>
+<script src="./js/login.js"></script>
